Fixes #53, jump, fall, and idle animations for all swords except guitar.
The sword animations have their own controllers for the slashing, but the idle is handled by the parent object's controller (Aka the player). This means the Player PreFab now needs the swords saved with it. So when we switch swords, we need to disable the current and enable the selected sword instead of the temp-hacky tab key version we currently have